Working login and registration

+++ user/models.py
@@ -5,7 +5,7 @@ from datetime import datetime
 # Create your models here.
 class Student(models.Model):
     user = models.OneToOneField(User)
-    mobile = models.IntegerField()
+    mobile = models.BigIntegerField()
     newsletter = models.BooleanField(default=0)
     book_count = models.IntegerField(default=0)
     last_visited = models.DateTimeField(default=datetime.now)
